feat: Initial steps to extract the 3 Penny-Arcade comic panels.
As mentioned in the docstring, Penny-Arcade doesn't put the comic link into the
RSS feed entry, so we need to explicitly load the link and then parse the full
page for the comic panels.

Penny-Arcade has also broken their comic on each panel to support
vertical/horizontal rendering.

This commit adds a function that can do the page load and extraction of comic
URL's.